requirements.txt
Flask
selenium
beautifulsoup4
pandas
numpy
mysql-connector-python